backend/requirements.txt
Django==3.0.3
djangorestframework==3.11.0
pynamodb==5.0.3
pynamodb-attributes==0.3.0
